Add SongManager::ChooseRandomSong (from TitleMenu).  Use it when starting a demonstration.
Instead of queuing a bunch of SM_PlayAttract's, queue one and
have it requeue itself.

This means 1: if the demonstration fails to start, we don't stop
playing the attract.  (Well, if we don't have any songs, maybe
we don't want ot attract attention, but ...) and more importantly,
2: setting the SECONDS_BETWEEN_ATTRACT value to some obscenely
high value won't cause millions of messages to be queued.

Flatten the option selection loop a bit (this should probably be
moved somewhere, too.)


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@228 822b03c2-8f45-0410-a444-95660a741a6f