Add Disasm to the projects.
Add build rules for incrementing the build number.  (Note that since
VC6 doesn't have pre-build rules, we have to compile and link it manually.
It's a somewhat ugly hack, but it works.  In VC7 we can just make it a
normal source file and regenerate it as a pre-build step.)

Turn on .MAP generation.  Make sure it always goes into the intermed
path, so it doesn't clutter the output path.

Add rules for generating the VDI debug file from .MAPs.

(I'm not committing the code to actually enable this yet; I want to iron
out any problems these build rules cause first.)


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@413 822b03c2-8f45-0410-a444-95660a741a6f