Clean up and optimize splitrelpath.
Add GetExtDirListing and GetExtDirListingV, for fast searches for
files by directory, filename prefix and extension.

Note a change in behavior due to this: all theme files will load as
prefixes, not just graphics.  (eg. "title sound (abc)" will load for "title sound",
just like "title graphic (abc)" loads for "title graphic".)  This seems more
consistent anyway, but if it's a problem it's easily fixed.

(Profiling showed that we were spending a decent amount of time reading
directories; this is because we were making separate calls for each
theme extension.)


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@461 822b03c2-8f45-0410-a444-95660a741a6f