Add new files.
Turn off browse info in vc7.  I know this takes a long time in 6; it
might not help much in 7.  I'll get around to timing it and turn it back on
if it doesn't matter.

Boost the warning level in vc7.  Not all of these warnings have been
cleaned up, but I'll get to them.


git-svn-id: https://svn.stepmania.com/svn/trunk@609 822b03c2-8f45-0410-a444-95660a741a6f