Fix SetPositionSamples() when we have a rate.  This should fix the editor problems.
There's a major restriction of rate changing: you can't change the
rate while the song is playing.  That's because RageSound keeps
track of the number of samples actually *output to the sound driver*;
it determines the position within the sound by dividing by the rate.
(It might be possible to adjust position dynamically within SetPlaybackRate
to make this work, but this isn't done right now.)


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@2521 822b03c2-8f45-0410-a444-95660a741a6f