Start loading the gameplay screen as soon as the stage graphic is finished, which reduces unnecessary delay.
This also means that it's possible to have, say, an Extra stage screen with
a long, complicated graphic, without having to make every screen delay
that long before prepping.

(It's currently not possible to have different switch delays for each type; eg.
Stage 1 prepping in 1 second and switching in 3, and Extra prepping in 3
seconds and switching in 5; the switch time is constant, based on m_In.
This would be useful in the longer-complicated-extra-stage case.)


git-svn-id: https://svn.stepmania.com/svn/trunk@2663 822b03c2-8f45-0410-a444-95660a741a6f