Fix UseSongBG
Hmm.  Need some way of using rendering masks in BGAs.  Not full
access to the zbuffer, of course, just a way to implement things like
the masked-off stage scroll on top of an unmasked song background.

Being able to do this based on the alpha values of another image would
have a whole lot of bonuses for background animations, too; eg the
canonical trick of rendering things masked by a silhouette.  This is tricky
to implement, though.

It'd be even better if we could do this with full alpha.

Most multipass algorithms I can think of would require an alpha
channel on the framebuffer, though, and we very often don't have
that.  I'll keep thinking about it (and perhaps randomly brainstorming
into the nearest edit window, such as WinCVS ...)


git-svn-id: https://svn.stepmania.com/svn/trunk@2664 822b03c2-8f45-0410-a444-95660a741a6f