Implement textures that only have an alpha channel.
The most immediate use of this is for the company screen shadow;
it's horribly banded in 16-bit, and this lets us load it as an 8-bit texture,
using half the memory and giving full precision.

This could be automatically used, if all non-transparent pixels in a texture
are #FFFFFF.  (todo)

This can also allow us to have extremely large, memory-compact fonts.
A 1024x1024 layer in ALPHA4 is just 512k, and can store several thousand
16x16 glyphs.  (The only trouble is actually constructing such a font;
doing it by hand in Photoshop isn't a nice option.)


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@2748 822b03c2-8f45-0410-a444-95660a741a6f