Add TimingAssist; this roughly implements the timing helper in PNM.
This lets us add an announcer that's played after each non-perfect
step, eg. "gameplay assist late good".

It only works reasonably for sparse songs--songs that only have notes
far enough apart that the announcer makes sense.  That's mostly
beginner songs, which is what this is targetted at anyway.  It's meaningless
to use this assist mode in harder songs; even if the announcer was
rate limited, it'd be impossible to tell which step the announcer was for,
so I didn't bother trying.  (It'd be amusing to turn this on in MaxU. :)

A different timing assist type (TimingTip) is probably coming soon,
targetted to higher level (light/med) songs.

The only major problem at the moment is that there's no way to tell
who the announcer is talking to if there's two players.  If we implement
per-player announcers, that would help (and allow a number of other
neat things, like having the p1 announcer say "Ready" and the p2
announcer say "HWG").


git-svn-id: https://svn.stepmania.com/svn/trunk@2784 822b03c2-8f45-0410-a444-95660a741a6f