Add an option to force preloading.  This only preloads sounds that are loaded in advance, which excludes sample music and some announcers.
Disable "sound volume".  That's not what that variable is for; setting
it to higher values will result in clipping on a lot of hardware, causing
people to report it as a "bug" (which it's not).  Setting it to lower values
results in lower quality sound.  To do this correctly, we'd need to add
an interface for accessing the real hardware mixer volume, which I
don't really want to do right now ...


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@2859 822b03c2-8f45-0410-a444-95660a741a6f