Use mySDL_EventState.  Set up event types in the ctor.
Restore event masks when we change modes.  It's vital that we not
do this incorrectly.  Resetting the video system resets the mask.  If we
have events that are enabled that we don't actually monitor, it'll just stay
in the event queue and eventually fill it.

Re-fix alt-enter causing both keys to "stick".


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@3436 822b03c2-8f45-0410-a444-95660a741a6f