Player::Update: Don't round for CrossedRow + explanation.
GetClosestNoteDirectional: Always round up, so we always look far enough.
It's harmless to look a little too far.  Fixes not being able to hit a note .5
of a note row too soon--there was small window in which you hadn't been
graded and should have been able to register a boo, but steps were
ignored.  (This showed up in MAX3.)


git-svn-id: https://svn.stepmania.com/svn/trunk@3640 822b03c2-8f45-0410-a444-95660a741a6f