Refactor:
It seems to be no big deal to have a lot of small files; I think it's the
actual scan of the song tree that's expensive.  Simplify with this in
mind.

Load song banners when the song is loaded, so we don't load banners
for songs that aren't there.

Avoid touching the original banner when we're simply loading.


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@3664 822b03c2-8f45-0410-a444-95660a741a6f