This is a fix for: [ 806799 ] Number of songs to be played is displayed incorrectly.
The ScreenSelectStyle Info graphics are modified to omit the "Max Stage 3"  text on the info graphics, since it cannot change when people change their system settings.  Beyond that, it isn't true for Nonstop and Endless.

(This is my first time revising anything, if I am steeping on anyones toes I'd like to know how to work with this better)

-Chris Pitman


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@5724 822b03c2-8f45-0410-a444-95660a741a6f