I have modified SM files to add network support. I tested this under linux AND windows. There should be no issues, if there is, feel free to use AIM to IM me at CNLohr, or email me, webmaster@cnlohr.com.
I modified the .dsp and Makefile.am only to add
networksyncmanager and ezsockets.

I added some startup code to stepmania.cpp,
so that if the user does not desire network support
none will be enabled.

Scorekeeper and screengameplay are modified
so that information can be sent to the server as
well as allowing the game sync to work.

Thanks.


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@9045 822b03c2-8f45-0410-a444-95660a741a6f