Added --Course command, and a command to send all songs to server so the server can index them for gameplay for SMOnline.
I am not sure if I did the Course setup properly.  If I did not, feel free to fix
it or... if you don't feel like it, I'm sure I'll figure it out after a while.

I left a lot of notes in... I am not sure what style of code you want,
at what point you want me to submit, if I am doing any of this wrong
please just tell me, and I'll try to be more useful.


git-svn-id: https://svn.stepmania.com/svn/trunk@9602 822b03c2-8f45-0410-a444-95660a741a6f