If a drive isn't ready, don't try to mount it at all; it may cause mkdir() calls to be made (if eg. mounting "D:\foo"), causing "volume not ready" dialogs.  This happens with removable devices like memory card readers that don't have a device.
I think this is generally incorrect: if a user has a memcard reader, we'll ignore
it on the first change (no device is installed), and then we'll never actually
notice when a card is inserted, since the drive letters don't actually
change.  I don't have a memcard reader to test this, so I'm punting on
this for now.


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@9782 822b03c2-8f45-0410-a444-95660a741a6f