Merge duplicate sounds.
Moving sounds which are used in various places to a separate common
name, and rediring to it, instead of picking an arbitrary element to
store the ogg and having everything else point at it.  This makes
changing them individually in other themes easier, and makes changes
to one screen less likely to cascade and require changes to others.

I'm only actually doing this now because I want to try to fix some sound
quality issues (at least partially due to previous reencoding and resampling),
and having lots of duplicate sounds was a pain.

_common value: EditMenu value, JukeboxMenu value, ScreenMiniMenu value
_common row: EditMenu row, JukeboxMenu row, ScreenEdit line, ScreenMiniMenu row, ScreenPrompt change
Common start: ScreenNameEntry step
_common change: ScreenSelectGroup change, ScreenSelectStyle change


git-svn-id: https://svn.stepmania.com/svn/trunk@11531 822b03c2-8f45-0410-a444-95660a741a6f