Modify AutoScreenMessage to keep map of all non-standard ScreenMessages.
This has a lot of advantages over the old ways.  Some of the initial ones are that
you can log a list of all screenmessages and their respective number and another
is that it would be fairly easy to add a reverse lookup, so you could get the name
of a ScreenMessage for log use by getting the ScreenMessage number.


git-svn-id: https://svn.stepmania.com/svn/trunk@14788 822b03c2-8f45-0410-a444-95660a741a6f