FEATURE:  Allow BitmapTexts to on-the-fly change color mid-text depending on the text
This feature can be used by enabling the coloration by using the .SetDynamicColor( true )
function.  Once this is done, the syntax for color changing is: |c0RRGGBB where RR = red
in hex, GG = green in hex and BB = blue in hex.  |c0 is the token to note the
beginning of a text change.  This was discussed on the StepMania development
list.


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@15402 822b03c2-8f45-0410-a444-95660a741a6f