Revert NotesLoaderBMS change.  It broke BMS loading (at least in 4-panel Dance mode).
The code is correct.  BMS_ values don't line up to any particular track
values.  The mapping given by mapBMSTrackToDanceNote simply maps the
weird track values in BMS files to consecutive values.  This means that
we'll gracefully handle unexpected values; for example, data on track
18 in 4-panel Dance mode.  We then transform the resulting data according
to the game type with NoteData::LoadTransformed.  All that matters is that
the BMS values used in the tables (switch) for the transformation match up
with those given by mapBMSTrackToDanceNote; the fact that BMS_ values
are 1-based and not 0-based doesn't matter.

The problem is probably that BMS values exceed MAX_NOTE_TRACKS: BMS_P2_KEY7
is 16, and MAX_NOTE_TRACKS was 16 (the greatest valid track value was 15).
Added more TRACK_ values.

This is tested and works with STEPS_TYPE_DANCE_SINGLE.  If it doesn't work
with wider data, it's probably not very important; 3.9 doesn't have enough
support to make most of those game types useful anyway ...


git-svn-id: https://svn.stepmania.com/svn/branches/v390@15620 822b03c2-8f45-0410-a444-95660a741a6f