We've been trying to more or less keep note skin compatibility, but the format has gone somewhat out of hand, so let's give up a bit on that for the good of mankind ...
All note elements that are allowed to be note colored *always* have a note
type in the filename; if XIsNoteColor is off, "4th" is used.  This way, we
don't have different sets of filename rules, fixing clashes between skins
and the fallback, and a number of things become simpler (many of which
I'm still discovering, so I won't try to enumerate them here just now).

Note skin updates coming ...


git-svn-id: https://svn.stepmania.com/svn/trunk@16461 822b03c2-8f45-0410-a444-95660a741a6f