fix course mode.
We've been using the environment and Lua globals as "parameters" to treat
broadcasts as "functions called on every object".  That doesn't work with
queued message handling.  It's also a bit confusing; it means there are
"parameters" available only for some messages.  It's cleaner to just make the
information available as a function.


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@16599 822b03c2-8f45-0410-a444-95660a741a6f