"XToThemedString" uses theme metrics.  We don't want DateTime (used by XmlFile) to depend on THEME, so use localized strings.
We may *always* want LocalizedString.  One reason we may not
is if we start using the DynamicThemeMetric paradigm more.
That evaluates expressions when they're used, which allows
metrics to change dynamically.  However, we may really not
want to do that for localized strings, since it would complicate
translation, so we may want to restrict localized strings to
pre-evaluation anyway.


git-svn-id: https://svn.stepmania.com/svn/trunk@19204 822b03c2-8f45-0410-a444-95660a741a6f