We have two "gamestate Lua environments": mapEnv and m_Environment.  The former is a map<CString,CString>; the latter is a Lua table.  The former is reset on gamestate reset, the latter is not.
Merge the map into the table, and eliminate it.  We only need one, and
it makes sense for a Lua environment to be a Lua table.

Do reset the table on Reset.  Otherwise, there's little point; if you want
to store data inside Lua, you don't need GameState to help.


git-svn-id: https://svn.stepmania.com/svn/trunk@19415 822b03c2-8f45-0410-a444-95660a741a6f