ScreenType is being overloaded in incompatible ways:  - the editor sets "system_menu" to trick StepMania.cpp into not    allowing the operator menu button, to prevent losing data;  - the editor also sets "gameplay" in "playback" mode to influence    ScreenSyncOverlay and ScreenDebugOverlay; this breaks the above,    allowing scroll lock in play mode.
A similar problem happens when we push screens: scroll lock works
in the MiniMenus in the editor.

It's semantically unintuitive to call the editor a "system_menu", anyway.

Split out AllowOperatorMenuButton, and use it directly.  Check all
screens, and don't allow it if any screen on the stack disallows it.


git-svn-id: https://svn.stepmania.com/svn/trunk@20125 822b03c2-8f45-0410-a444-95660a741a6f