Add parent links to actors.
For each actor, create a context table, with an __index link to the
parent's.  This can be used to propagate information down complex
actor trees.

Code needs to be changed to do AddChild() calls before loading,
rather than after, so context links are available during loads.
This will be done as necessary.


git-svn-id: https://svn.stepmania.com/svn/trunk@22105 822b03c2-8f45-0410-a444-95660a741a6f