Ideally, actors should be standalone: instanced and maintained in XML without the screen knowing about it.  That's not always feasible, for actors with more complicated interactions, such as SSMusic's MusicWheel, Banner, and so on.
However, if we instance those from the screen itself, then it's pulled outside
of the standard XML heirarchy: it has to go around the major layers (overlay,
underlay).  That's a pain.

Create ActorProxy.  This simply points to and draws another actor.  To use,
hide the main actor (hidden,1), create an ActorProxy in XML, and point it
at the actor you want.  The actor is still controlled by and owned by the screen,
but can be layered in any place, and instanced multiple times.

(This does not propagate things like ActorFrame; that would cause the
target actor to be updated and receive commands multiple times, once
for each proxy.  The owner, and path for those calls, is still the screen.)


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@22216 822b03c2-8f45-0410-a444-95660a741a6f