Previously, changing themes attempted to preserve Lua data by serializing data and tables while reinstantiating Lua.  That's ugly, and can't serialize some things (eg. functions).
Instead, when changing themes, shut down the high-level objects
and unregister as much as we can from Lua, change the theme,
then reregister things.  This doesn't clean the slate as much;
globals won't be erased, etc.

A better way would be to clear everything that uses Lua
(delete singletons), then reload it all.  Singleton creation is
a bit of a mess currently so that would break a lot, and we
don't want to reload songs.


git-svn-id: https://svn.stepmania.com/svn/trunk@22867 822b03c2-8f45-0410-a444-95660a741a6f