I'm not sure what was happening here:
If ALLOW_DISABLED_PLAYER_INPUT is true (only ScreenTitleMenu), then
ScreenSelect::Input doesn't do JoinInput ("the player will be joined
and the user will still be on the title menu").  However,
ScreenSelectMaster::MenuStart does the JoinInput anyway, so this still
happens.

If we want to have main menu options that don't load a screen, this
code path should be refactored a bit, to ask the derived class whether
the selection should also join.  As is, it's just making the JoinInput
path complicated.


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@23991 822b03c2-8f45-0410-a444-95660a741a6f